Forking Python uiautomator interface into devtest.
Due to extensive modifications that need to be made to it.
Since it's MIT licensed it should be fine.

Ripped out its adb support and replaced it with devtest's.
Remove subprocess dependency (and remove associated bug).
Added the existing Google fixes.
Add support for switching between google and upstream server side
(for comparison).