Ignore modules that have "analyze" in the name.
This enables peer-modules that analyze data that test cases can write
out. This data has user-defined structure, thus making production (in
test cases) and analysis of it tightly coupled. At the same time, we
want to reduce test case and framework dependencies. Analysis modules
typically have additional dependencies that are quite heavy, that the
test case doesn't need.

So, the intent is to write peer-modules that live in the same packaage
as the test case but are otherwise decoupled from it. They should have
"analyze" in the name. So you have:

    mytests.py
    mytests_analyze.py

For tests that produce data that is persistent, and can be analyzed.

Data is serialized to JSON by the framework. So the analysis module has
to know how to convert some of the data back to Python objects.